Prevent a Super Admin user from being deleted

+++ app/Services/UserService.php
@@ -279,6 +279,9 @@ class UserService extends BaseService {
 	 */
 	public function delete(User $user) : Response
 	{
+		if ($user->isSuperAdmin()) {
+			return Response::error(__('You cannot delete a Super User'), 408);
+		}
 		$this->repository->delete($user);
 		return Response::success();
 	}
